extract course tests
